online_inference/requirements.txt
scikit-learn~=1.0.2
pandas~=1.4.2
pydantic~=1.9.1
uvicorn~=0.17.6
fastapi~=0.78.0
joblib~=1.1.0
python-dotenv~=0.20.0
lightgbm==3.3.2
